Extract LocalhostGuardConfigSupport shared by MVC and Reactive localhost filters
LocalhostOnlyFilter (servlet) and ReactiveLocalhostOnlyFilter (WebFlux) each
carried a byte-identical ~90-line copy of the guard-config-building logic:
buildConfig(), toGatewayTrust(), the lazy cached container-gateway resolution,
the trusted-proxies CIDR parse cache, and the once-only "trusting container
gateway" warning. None of that logic touches a servlet or reactive request
type - it only translates BootUiProperties (+ ContainerGatewayDetector) into
a LocalhostGuardConfig - so it is extracted into a new
LocalhostGuardConfigSupport class that both filters construct and delegate
to.

Logger is threaded through as an explicit parameter (buildConfig(Logger),
warnTrustedGatewayOnce(Logger, InetAddress)) rather than giving the shared
class its own static logger, so log lines keep attributing to the concrete
filter class exactly as before. Each filter still owns its own
remoteAddress()/logRejection() glue, which does read transport-specific
request types and is appropriately NOT shared.

Updated both filters' class javadoc to describe the new shared-helper
reality instead of the old "duplicates intentionally, mirroring the Quarkus
independent translation layer" wording (which never applied to this
particular logic - it doesn't touch servlet or reactive types at all.

Verified: LocalhostOnlyFilterTests (35) and ReactiveLocalhostOnlyFilterTests
(16) pass unchanged; full bootui-core/-engine/-spring-autoconfigure suite
1267 tests, only the pre-existing unrelated flaky
ReactiveBootUiChangeStreamTests timeout (confirmed passes in isolation);
spotless:apply clean.

EOF
)
